feat: Connect ball object to PonGame widget
So I guess similar to what I learned in Udacity's intro to Android,
the kivy language plays the role of XML and python plays the role
of Java. In this situation, in the kv rules file pong.kv, we add
the rule in PongGame that PongBall objects will be refered to by an
id of pong_ball. Since the PongGame class in main.py has a *ball*
variable initiated with ObjectProperty from kivy.properties, the rule
in pong.kv for <PongGame>;

ball: pong_ball

will let the app know it is actually of type PongBall.