feat: Add PongPaddle class which is player class
Began writing logic for two players each which is an instance of
the paddle class. PongPaddle houses score variable which is a
NumericProperty. As for pong.kv, the <PongPadle> descriptor is
simlar to <PongBall> basically a widget containing a self centered
rectangle of its canvas dimensions. Now, two players are instantiated
in main.py in PongGame as ObjectProperty, they are hooked to PongPaddle
type in pong.kv by adding two PongPaddle widget intries in <PongGame>.
Each with id corresponding to left and right players.